Refactored and built for more robust measuring

+++ functions/bandpassFiltData.m
@@ -7,6 +7,6 @@ function filteredData = bandpassFiltData(data, lowhigh, order)
 
 	[b,a] = ellip(order, .2, 60, lowhigh, 'bandpass');
 
-	filteredData = filtfilt(b,a, data);
+	filteredData = filtfilt(b,a, double(data));
 
 end
